requirements.txt
requests
json
logging
pyyaml
assertpy
pytest
flake8
pylint
pycodestyle
pytest_bdd
sttable
jsonschema